Move Sass render “all component” tests to `components.test.js`
These tests don’t need a Chromium browser window (Puppeteer environment)
